# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserProfile": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"aboutMe": {
						"type": "string"
					},
					"phoneCountryCode": {
						"type": "number",
						"format": "double"
					},
					"phoneNumber": {
						"type": "number",
						"format": "double"
					},
					"yearOfBirth": {
						"type": "number",
						"format": "double"
					},
					"location": {
						"type": "string"
					},
					"userType": {
						"type": "string"
					},
					"themeColor": {
						"type": "string"
					},
					"spotifyPlaylistUrl": {
						"type": "string"
					},
					"onboardingCompleted": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"firstName",
					"lastName",
					"location",
					"userType",
					"onboardingCompleted",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FeedPostResponse": {
				"description": "API response format for a post in the feed\r\nExtends PostResponse with engagement data and user interaction state",
				"properties": {
					"likesCount": {
						"type": "number",
						"format": "double"
					},
					"commentsCount": {
						"type": "number",
						"format": "double"
					},
					"bookmarksCount": {
						"type": "number",
						"format": "double"
					},
					"hasLiked": {
						"type": "boolean"
					},
					"hasBookmarked": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ConnectionStatus": {
				"type": "string",
				"enum": [
					"pending",
					"accepted",
					"rejected"
				]
			},
			"Connection": {
				"properties": {
					"id": {
						"type": "string"
					},
					"requesterId": {
						"type": "string"
					},
					"recipientId": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ConnectionStatus"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"requester": {
						"$ref": "#/components/schemas/UserProfile"
					},
					"recipient": {
						"$ref": "#/components/schemas/UserProfile"
					}
				},
				"required": [
					"id",
					"requesterId",
					"recipientId",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateConnectionRequestDto": {
				"description": "DTO for creating a connection request\r\n\r\nUsed when sending a connection request to another user.\r\nThe requester_id is automatically extracted from the authentication token.",
				"properties": {
					"recipient_id": {
						"type": "string",
						"description": "The ID of the user to send the connection request to (UUID format)",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					}
				},
				"required": [
					"recipient_id"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"recipient_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
				}
			},
			"UpdateConnectionRequestDto": {
				"description": "DTO for updating a connection request status\r\n\r\nUsed when accepting or rejecting a connection request.\r\nOnly the recipient can update the status.",
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ConnectionStatus",
						"description": "The new status for the connection request",
						"example": "accepted"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": "accepted"
				}
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "LineUp API",
		"version": "1.0.0",
		"description": "API documentation for the LineUp application",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Pinnaacle"
		}
	},
	"paths": {
...
		"/feed": {
			"get": {
				"operationId": "GetFeed",
				"responses": {
					"200": {
						"description": "Feed posts with pagination cursor, or recommendations array",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"nextCursor": {
													"type": "string"
												},
												"posts": {
													"items": {
														"$ref": "#/components/schemas/FeedPostResponse"
													},
													"type": "array"
												}
											},
											"required": [
												"posts"
											],
											"type": "object"
										},
										{
											"properties": {
												"recommendations": {
													"items": {
														"$ref": "#/components/schemas/FeedPostResponse"
													},
													"type": "array"
												}
											},
											"required": [
												"recommendations"
											],
											"type": "object"
										},
										{
											"properties": {
												"stories": {
													"items": {},
													"type": "array"
												}
											},
											"required": [
												"stories"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Get home feed\r\n\r\nReturns feed content based on the specified component:\r\n- \"posts\": Posts from users you follow (with pagination)\r\n- \"recommendations\": Collaboration request recommendations\r\n- \"stories\": Stories from followed users (coming soon)\r\n\r\nPosts feed includes engagement data (likes, comments, bookmarks counts)\r\nand user interaction state (hasLiked, hasBookmarked).",
				"summary": "Get home feed",
				"tags": [
					"Feed"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The feed component to fetch (default: \"posts\")",
						"in": "query",
						"name": "component",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"posts",
								"recommendations",
								"stories"
							]
						}
					},
					{
						"description": "Cursor for pagination (ISO timestamp)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Maximum number of items to return (1-50, default: 20)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/connections": {
			"get": {
				"operationId": "GetConnectionRequests",
				"responses": {
					"200": {
						"description": "Array of connection requests (sent and received)",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Connection"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all connection requests for the authenticated user\r\n\r\nReturns both sent and received connection requests for the authenticated user.\r\nEach request includes requester and recipient profile information.",
				"summary": "Get user's connection requests",
				"tags": [
					"Connections"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateConnectionRequest",
				"responses": {
					"200": {
						"description": "The created connection request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Connection"
								}
							}
						}
					}
				},
				"description": "Create a connection request\r\n\r\nAllows the authenticated user to send a connection request to another user.\r\nThe recipient must exist and there must not be an existing connection request.",
				"summary": "Send a connection request",
				"tags": [
					"Connections"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Connection request data including recipient ID",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateConnectionRequestDto",
								"description": "Connection request data including recipient ID"
							}
						}
					}
				}
			}
		},
		"/connections/sent/{userId}": {
			"get": {
				"operationId": "GetSentConnectionRequests",
				"responses": {
					"200": {
						"description": "Array of connection requests sent by the user",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Connection"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get connection requests sent by a user\r\n\r\nReturns all connection requests sent by the specified user.\r\nThis endpoint can be accessed without authentication.",
				"summary": "Get sent connection requests by user",
				"tags": [
					"Connections"
				],
				"security": [],
				"parameters": [
					{
						"description": "The UUID of the user to retrieve sent requests for",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/connections/received/{userId}": {
			"get": {
				"operationId": "GetReceivedConnectionRequests",
				"responses": {
					"200": {
						"description": "Array of connection requests received by the user",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Connection"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get connection requests received by a user\r\n\r\nReturns all connection requests received by the specified user.\r\nThis endpoint can be accessed without authentication.",
				"summary": "Get received connection requests by user",
				"tags": [
					"Connections"
				],
				"security": [],
				"parameters": [
					{
						"description": "The UUID of the user to retrieve received requests for",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/connections/{requestId}": {
			"put": {
				"operationId": "UpdateConnectionRequest",
				"responses": {
					"200": {
						"description": "The updated connection request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Connection"
								}
							}
						}
					}
				},
				"description": "Update a connection request status\r\n\r\nAllows the recipient to accept or reject a connection request.\r\nOnly the recipient can update the status, and only pending requests can be updated.",
				"summary": "Accept or reject a connection request",
				"tags": [
					"Connections"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The UUID of the connection request to update",
						"in": "path",
						"name": "requestId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Updated status (accepted or rejected)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateConnectionRequestDto",
								"description": "Updated status (accepted or rejected)"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteConnectionRequest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete a connection request\r\n\r\nRemoves a connection request. Only the requester can delete a pending request.",
				"summary": "Delete a connection request",
				"tags": [
					"Connections"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The UUID of the connection request to delete",
						"in": "path",
						"name": "requestId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}